move font to layout
