+++ components/Layout.js
@@ -1,4 +1,6 @@
 import Head from 'next/head'
+/* eslint-disable camelcase */
+import { Noto_Sans } from 'next/font/google'
 import Nav from './Nav'
 import Banner from './Banner'
 import EditBanner from './EditBanner'
@@ -7,6 +9,14 @@ import Footer from './Footer'
 import FooterMinimal from './FooterMinimal'
 import FeedbackModal from './FeedbackModal'
 import BibtexModal from './BibtexModal'
+
+const notoSans = Noto_Sans({
+  subsets: ['latin'],
+  weight: ['400', '700'],
+  style: ['normal', 'italic'],
+  display: 'fallback',
+  variable: 'test',
+})
 
 export default function Layout({
   children,
@@ -18,7 +28,7 @@ export default function Layout({
   minimalFooter,
 }) {
   return (
-    <>
+    <div className={notoSans.className}>
       <Head>
         <meta charSet="utf-8" />
         <meta name="viewport" content="width=device-width, initial-scale=1" />
@@ -91,6 +101,6 @@ gtag('config', '${process.env.GA_PROPERTY_ID}', {
 
       <FeedbackModal />
       <BibtexModal />
-    </>
+    </div>
   )
 }
