+++ styles/global.scss
@@ -11,7 +11,7 @@ html {
   }
 }
 body {
-  font-family: 'Noto Sans', 'Helvetica Neue', Helvetica, Arial, sans-serif;
+  font-family: 'Helvetica Neue', Helvetica, Arial, sans-serif;
   font-size: 0.875rem;
 }
 #__next {
